Pre-commit lint fixes
